[WIP]: Sections

+++ src/controllers/quiz.js
@@ -14,6 +14,9 @@ import {
 } from '../models/quiz';
 
 const controller = {
+	/**
+	 * Returns the list of all quizzes only for superadmins
+	 */
 	getAllQuizzes: async (req, res) => {
 		const quizzes = await getAllQuizzes();
 		if (quizzes) {
@@ -21,7 +24,10 @@ const controller = {
 		}
 		return notFoundResponse(res);
 	},
-
+	/**
+	 * Returns the requested quiz data only for superadmins,
+	 * quiz creator, quiz owners and quiz registrants
+	 */
 	getQuizById: async (req, res) => {
 		const { username, role } = req.user;
 		const { quizioID } = req.params;
@@ -38,7 +44,9 @@ const controller = {
 
 		return notFoundResponse(res, 'Quiz not found!');
 	},
-
+	/**
+	 * A new quiz is added to the db with the requesting user as the creator
+	 */
 	addNewQuiz: async (req, res) => {
 		const { username } = req.user;
 		const quiz = await addNewQuiz(username);
@@ -50,7 +58,10 @@ const controller = {
 		}
 		return errorResponse(res, 'Failed to add new quiz!');
 	},
-
+	/**
+	 * Updates the quiz to the data sent in the body only when superadmin
+	 * or quiz owner or quiz creator makes the call
+	 */
 	updateQuiz: async (req, res) => {
 		const { username, role } = req.user;
 		const { quizioID } = req.params;
@@ -70,14 +81,25 @@ const controller = {
 		}
 		return notFoundResponse(res, 'Quiz not found!');
 	},
-
+	/**
+	 * Deletes the quiz only when superadmin
+	 * or quiz owner or quiz creator makes the call
+	 */
 	deleteQuiz: async (req, res) => {
+		const { username, role } = req.user;
 		const { quizioID } = req.params;
-		const deleted = await deleteQuiz(quizioID);
-		if (deleted) {
-			return successResponseWithMessage(res, 'Quiz deleted successfully!');
+		const quiz = await getQuizById(quizioID);
+
+		if (quiz) {
+			if (role === 'superadmin' || quiz.creator === username || quiz.owners.includes(username)) {
+				const deleted = await deleteQuiz(quizioID);
+				if (deleted) {
+					return successResponseWithMessage(res, 'Quiz deleted successfully!');
+				}
+			}
+			return unauthorizedResponse(res);
 		}
-		return errorResponse(res, 'Failed to delete quiz!');
+		return notFoundResponse(res, 'Quiz not found!');
 	},
 
 };
